Fix storage.size() for gzipped content via custom metadata
This does not work if the file was uploaded from other sources/libs
as the custom metadata will not be there. In this case there is no
cheap way to get the _unzipped_ file size without downloading it.

It _may_ be possible to download just the "gzip metadata" part of
the file and get the size info from there. Such possibility is
let as an exercise for the developer. If someone accomplish that,
please send us as a Pull-Request :)